test: move test_collect_interrupted to test_dataframe.py
